[close GH-1] Move core logic from class into a module
By moving the functionality from a class to a module we can use Wicked in classes that inherit from things other the ApplicationController, this is especially useful if we have a top level controller such as AdminController, or ApiController that inherits from ApplicationController and we wish to inherit from it. You should use the include method rather than inheriting directly.

Does not break backwards compatibility