Prevent Symbol DoS Attacks
Prior to this Commit, any Website utilizing Wicked is susceptible to a DoS attack. Ruby does not garbage collect symbols (:first, :last, :user, etc.) so every symbol you allocate will stay in memory for the life of the process. This can be a problem if an attacker is able to create a very large number of symbols in a system. Prior versions of wicked called params[:id].to_sym which will create a symbol for every new string passed into :id, this could allow an attacker to send enough symbols to crash a running instance. 

This commit fixes this vulnerability by using values popped off of `steps` that match the params[:id] after both have been turned into strings. This preserves the data-type of the steps (symbols versus strings).