add failing tests for translation helpers
In issue #51 @seizmo pointed out that there is an inconsistency with the step name and steps variables that are being used. These tests should pass if wicked was working correctly, unfortunately there is internal confusion over what the state of steps and step is i.e. have they been translated yet, or are they still the canonical values?