closes #51 fix step translation behavior
Rather than be clever by translating some values and not others, `step` and the values in `steps` will now always be the same language. This means that now all the step helpers work correctly, but if someone needs to use a canonical step value in their controller such as using `step` with a case statement they must now use `wizard_value()` to cohere the value into the canonical value like this:

```
case wizard_value(step)
when :first
  # …
when :second
  #…
```

wizard_value will always return a value in the initially declared steps for example if the locale is set to spanish and steps are set to:

```
steps :first, :second
```

Then the value `wizard_value()` will only return a value of `:first` or `:second`:

```
wizard_value(:uno) # => :first
```

If the language is set to english it will still return the correct value:

```
wizard_value(:first) # => :first
```

The important thing to remember is that `step` and the values in `steps` are always going to be in the same language if you're using the Wicked translations. If you need any values to match the values set directly in your controller, or the names of your files (i.e. `views/#../first.html.erb`, then you need to use `wizard_value` method.