Raise exceptions on invalid step requests
The current implementation accepts the params[:id] value without 
question, which leads to the potential for vulnerabilities like #94, 
since it allows potential attackers/clever users to bypass the step 
definitions established in the controller.

This performs a strict check of the requested step against the steps 
defined for the Wicked::Wizard controller, and raises an exception if
no matching step can be identified. For the sake of friendly error-
handling (e.g., rendering 404 pages appropriately), special runtime 
error classes have been introduced.