Remove jeweler/ fix tests
Running into a jruby error even when testing against MRI. This fixes the issue by only requiring jruby in the Gemfile where a platform option can be specified instead of in the gem spec.

Removing jeweler allows us to to be more flexible with our gemspec and it wasn't really all that needed.