[Close #159] Fix double params
The `index` action assumed that there are no params when generating a URL, however this is not always true.

Someone can dynamically inject params by overwriting `default_url_options` like so:

```ruby
  def default_url_options(options = {})
    { locale: I18n.locale }.merge options
  end
```

When they do this, the redirect code in `index` would cause double params to show up.

The fix is to pass in params directly instead of manually trying to build up a URL for redirect. The solution is cleaner as well.