Run CI on a recurring schedule
In addition to running on every push or pull request, the CI action
will now run on a recurring schedule of once a month.

The schedule is set to run on the 1st of each month at 02:03 UTC time.
The time is completely arbitrary but I avoided the top/bottom of the
hour since most cron jobs tend to occur there.